requirements.txt
-f https://download.pytorch.org/whl/nightly/cpu
-f https://iree.dev/pip-release-links.html
--pre

setuptools
wheel


torch==2.3.0
iree-turbine @ git+https://github.com/iree-org/iree-turbine.git@main
turbine-models @ git+https://github.com/nod-ai/AMDSHARK-ModelDev.git@main#subdirectory=models
diffusers @ git+https://github.com/nod-ai/diffusers@0.29.0.dev0-amdshark
brevitas @ git+https://github.com/Xilinx/brevitas.git@6695e8df7f6a2c7715b9ed69c4b78157376bb60b

# AMDSHARK Runner
tqdm

# AMDSHARK Downloader
google-cloud-storage

# Testing
pytest
Pillow
parameterized

# Add transformers, diffusers and scipy since it most commonly used
#accelerate is now required for diffusers import from ckpt.
accelerate
scipy
transformers==4.37.1
torchsde # Required for Stable Diffusion SDE schedulers.
ftfy
gradio==4.29.0
altair
omegaconf
# 0.3.2 doesn't have binaries for arm64
safetensors==0.3.1
py-cpuinfo
pydantic==2.4.1 # pin until pyinstaller-hooks-contrib works with beta versions
mpmath==1.3.0
optimum

# Keep PyInstaller at the end. Sometimes Windows Defender flags it but most folks can continue even if it errors
pefile
pyinstaller
